修改nk-checkbox-group类名
