+++ projects/ngx-uikit/src/components/checkbox/nk-checkbox-group.component.ts
@@ -1,22 +1,20 @@
-import {
-  ChangeDetectorRef,
-  Component,
-  forwardRef,
-  ViewEncapsulation
-} from '@angular/core';
-import {NG_VALUE_ACCESSOR} from '@angular/forms';
+import {ChangeDetectionStrategy, Component, EventEmitter, forwardRef, OnInit, Output, ViewEncapsulation} from '@angular/core';
+import {ControlValueAccessor, NG_VALUE_ACCESSOR} from '@angular/forms';
 import {NkCheckboxOption} from '../core/type/nk-types';
-import {isArray, isEmpty, isNil} from '../core/util/nk-util';
-import {NkBaseCheckbleGroupComponent} from '../radio/nk-base-checkble-group.component';
+import {NkCheckboxDirective} from './nk-checkbox.directive';
+import {isArray, isEmpty, isNil, isNotNil} from '../core/util/nk-util';
+import {BehaviorSubject} from 'rxjs';
+import {debounceTime, filter} from 'rxjs/operators';
 
 /**
- * 复选框组。 设计思想是,通过给定任意数组或Observable发射的数组创建多个复选框
- * 其中, label和value可以通过自定义属性名称或函数生成, 当ngModel
+ * nk-checkbox容器,可以访问子节点的nk-checkbox
+ *
  */
 @Component({
   selector: 'nk-checkbox-group',
-  templateUrl: './nk-checkbox-group.component.html',
+  template: '<ng-content></ng-content>',
   exportAs: 'nkCheckboxGroup',
+  changeDetection: ChangeDetectionStrategy.OnPush,
   encapsulation: ViewEncapsulation.None,
   providers: [
     {
@@ -26,91 +24,131 @@ import {NkBaseCheckbleGroupComponent} from '../radio/nk-base-checkble-group.comp
     }
   ]
 })
-export class NkCheckboxGroupComponent extends NkBaseCheckbleGroupComponent<NkCheckboxOption> {
+export class NkCheckboxGroupComponent implements OnInit, ControlValueAccessor {
   /**
    * ngModel绑定的值,该值通过valueFormat转换后获得
    * 如果列表为空,则设置为null,为了适应表单的required
    */
-  // tslint:disable-next-line
+    // tslint:disable-next-line
   modelValue: any[] | null = null;
 
-  constructor(public _cdr: ChangeDetectorRef) {
-    super(_cdr);
+  /**
+   * 触发ngModelChange的同时触发该事件,当所有复选框都没被选中时返回空数组
+   */
+  @Output() nkOnChange = new EventEmitter<NkCheckboxOption[]>();
+  /**
+   * 当前复选框状态改变时触发该事件
+   */
+  @Output() nkOnItemChange = new EventEmitter<NkCheckboxOption>();
+  /**
+   * 复选框点击事件
+   * 触发该事件时,并不能正确判断checked状态
+   * 要获取正确的checked状态,请使用nkItemChange
+   */
+  @Output() nkOnItemClick = new EventEmitter<{nkOption: NkCheckboxOption, clickEvent: MouseEvent}>();
+
+  /**
+   * 所有子节点的checkbox
+   */
+  _childrenNodes: NkCheckboxDirective[] = [];
+  _onChange = (_: object | null) => { };
+  _onTouched = () => { };
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+  registerOnChange(fn: (_: object | null) => { }): void {
+    this._onChange = fn;
+  }
+
+  registerOnTouched(fn: () => {}): void {
+    this._onTouched = fn;
   }
 
   // tslint:disable-next-line
-  toModelValue(value: any): any {
-    if (isEmpty(value)) {
-      return null;
+  writeValue(value: any[]): void {
+    this.modelValue = isEmpty(value) ? null : (isArray(value) ? value : [value]);
+    this._childrenNodes.forEach(child => child.modelValue = this.modelValue);
+    this.updateChildrenCheckedState();
+  }
+
+  setDisabledState(isDisabled: boolean): void {
+    this._childrenNodes.forEach(value => value.setDisabledState(isDisabled));
+  }
+
+  updateChildrenCheckedState(): void {
+    if (isEmpty(this._childrenNodes)) {
+      return ;
     }
 
-    return isArray(value) ? value : [value];
-  }
+    if (isNil(this.modelValue)) {
+      this._childrenNodes
+        .filter(ck => ck.getCheckedState())
+        .forEach(ck => {
+          ck.setCheckedState(false);
+        });
+      return ;
+    }
 
-  mapOptions(options: NkCheckboxOption[]): NkCheckboxOption[] {
-    return options;
-  }
-
-  /**
-   * 当个界面checkbox状态变化事件
-   * @param checked 是否被选中
-   * @param indeterminate 是否半选
-   * @param item 当前数据
-   */
-  // tslint:disable-next-line
-  itemViewChange(checked: boolean, indeterminate: boolean, item: NkCheckboxOption): void {
-    item.nkChecked = checked;
-    item.nkIndeterminate = indeterminate;
-    this.updateModelValue();
-    this.nkOnItemChange.emit(item);
+    this._childrenNodes.forEach(ck => ck.updateCheckedState());
   }
 
   /**
-   * 更新ngModel
+   * 更新ngModel的值
    * @param emit 是否发布ngModelChange事件
    */
   updateModelValue(emit: boolean = true): void {
-    const checkedItems = this._options.filter(val => val.nkChecked);
-    const checkedValues = checkedItems.map(val => val.nkValue);
-    this.modelValue = this.toModelValue(checkedValues);
+    const checkedItems = this._childrenNodes.filter(it => it.getCheckedState());
+    const checkedValues = checkedItems.map(it => it.nkValue);
+    this.modelValue = isEmpty(checkedValues) ? null : checkedValues;
+    const checkedOptions = checkedItems.map<NkCheckboxOption>(this._childToNkCheckboxOption);
     if (emit) {
       this._onChange(this.modelValue);
     }
-    this.nkOnChange.emit(checkedItems);
+    this.nkOnChange.emit(checkedOptions);
   }
 
   /**
-   * 获取用于检测是否选中的函数
+   * 子节点checkbox注册到该容
+   * @param child 节点checkbox
    */
-  // tslint:disable-next-line
-  protected getCheckFn(value: any | null): (_v: any) => boolean {
-    return isNil(value)
-      // tslint:disable-next-line
-      ? (_v: any) => false
-      // tslint:disable-next-line
-      : (_v: any) => (value as any[]).some(it => this.compareWith(_v, it));
-  }
+  registerChild(child: NkCheckboxDirective): void {
+    this._childrenNodes.push(child);
+    child.modelValue = this.modelValue;
+    child.updateCheckedState();
 
-  // tslint:disable-next-line
-  writeValue(obj: any[]): void {
-    this.updateCheckedState(obj);
-  }
-
-  /**
-   * 更新value
-   */
-  // tslint:disable-next-line
-  updateCheckedState(value: any | null): boolean {
-    this.modelValue = this.toModelValue(value);
-
-    if (isEmpty(this._options)) {
-      return false;
-    }
-
-    const fnChecked = this.getCheckFn(this.modelValue);
-    this._options.forEach(it => {
-      it.nkChecked = fnChecked(it.nkValue);
+    child.nkOnChange.subscribe(() => {
+      this.updateModelValue();
+      this.nkOnItemChange.emit(this._childToNkCheckboxOption(child));
     });
-    return true;
+  }
+
+  /**
+   * 把NkCheckboxDirective转换为NkCheckboxOption
+   * @param checkbox NkCheckboxDirective
+   */
+  private _childToNkCheckboxOption(checkbox: NkCheckboxDirective): NkCheckboxOption {
+    return {
+      nkValue: checkbox.nkValue,
+      nkChecked: checkbox.nkValue !== null,
+      nkLabel: checkbox.nkLabel,
+      nkDisabled: checkbox.nkDisabled,
+      nkIndeterminate: checkbox.nkIndeterminate
+    } as NkCheckboxOption;
+  }
+
+  /**
+   * 删除子节点的checkbox
+   * @param child 子节点的checkbox
+   */
+  removeChild(child: NkCheckboxDirective): boolean {
+    const index = this._childrenNodes.indexOf(child);
+    if (index > -1) {
+      this._childrenNodes.splice(index, 1);
+      return true;
+    }
+    return false;
   }
 }
