+++ projects/ngx-uikit/src/components/checkbox/nk-checkbox.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import {NkCheckboxDirective} from './nk-checkbox.directive';
-import {NkCheckboxGroupComponent} from './nk-checkbox-group.component';
-import { NkCheckboxContainerComponent } from './nk-checkbox-container.component';
+import {NkCheckboxListComponent} from './nk-checkbox-list.component';
+import { NkCheckboxGroupComponent } from './nk-checkbox-group.component';
 
-const COMPONENTS = [NkCheckboxDirective, NkCheckboxGroupComponent, NkCheckboxContainerComponent];
+const COMPONENTS = [NkCheckboxDirective, NkCheckboxListComponent, NkCheckboxGroupComponent];
 
 @NgModule({
   imports: [
