+++ projects/ngx-uikit/src/components/checkbox/public_api.ts
@@ -4,5 +4,5 @@
 
 export * from './nk-checkbox.module';
 export * from './nk-checkbox.directive';
+export * from './nk-checkbox-list.component';
 export * from './nk-checkbox-group.component';
-export * from './nk-checkbox-container.component';
